+++ app/src/main/java/app/revanced/manager/compose/di/RepositoryModule.kt
@@ -3,7 +3,8 @@ package app.revanced.manager.compose.di
 import app.revanced.manager.compose.domain.repository.ReVancedRepository
 import app.revanced.manager.compose.network.api.ManagerAPI
 import app.revanced.manager.compose.domain.repository.BundleRepository
-import app.revanced.manager.compose.domain.repository.SourceConfigRepository
+import app.revanced.manager.compose.domain.repository.SourcePersistenceRepository
+import app.revanced.manager.compose.domain.repository.SourceRepository
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
@@ -11,5 +12,6 @@ val repositoryModule = module {
     singleOf(::ReVancedRepository)
     singleOf(::ManagerAPI)
     singleOf(::BundleRepository)
-    singleOf(::SourceConfigRepository)
+    singleOf(::SourcePersistenceRepository)
+    singleOf(::SourceRepository)
 }
