+++ app/src/main/java/app/revanced/manager/compose/domain/repository/BundleRepository.kt
@@ -2,19 +2,19 @@ package app.revanced.manager.compose.domain.repository
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
-import app.revanced.manager.compose.domain.manager.patch.PatchBundle
+import app.revanced.manager.compose.patcher.patch.PatchBundle
 import app.revanced.manager.compose.util.launchAndRepeatWithViewLifecycle
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
 
-class BundleRepository(private val sourcesProvider: SourcesProvider) {
+class BundleRepository(private val sourceRepository: SourceRepository) {
     /**
      * A [Flow] that emits whenever the sources change.
      *
      * The outer flow emits whenever the sources configuration changes.
      * The inner flow emits whenever one of the bundles update.
      */
-    private val sourceUpdates = sourcesProvider.sources.map { sources ->
+    private val sourceUpdates = sourceRepository.sources.map { sources ->
         sources.map { (name, source) ->
             source.bundle.map { bundle ->
                 name to bundle
@@ -23,11 +23,16 @@ class BundleRepository(private val sourcesProvider: SourcesProvider) {
     }
 
     private val _bundles = MutableStateFlow<Map<String, PatchBundle>>(emptyMap())
+
+    /**
+     * A [Flow] that gives you all loaded [PatchBundle]s.
+     * This is only synced when the app is in the foreground.
+     */
     val bundles = _bundles.asStateFlow()
 
     fun onAppStart(lifecycleOwner: LifecycleOwner) {
         lifecycleOwner.lifecycleScope.launch {
-            sourcesProvider.loadSources()
+            sourceRepository.loadSources()
         }
 
         lifecycleOwner.launchAndRepeatWithViewLifecycle {
