+++ app/src/main/java/app/revanced/manager/compose/domain/repository/SourceRepository.kt
@@ -4,9 +4,9 @@ import android.app.Application
 import android.util.Log
 import app.revanced.manager.compose.data.room.sources.SourceEntity
 import app.revanced.manager.compose.data.room.sources.SourceLocation
-import app.revanced.manager.compose.domain.manager.sources.RemoteSource
-import app.revanced.manager.compose.domain.manager.sources.LocalSource
-import app.revanced.manager.compose.domain.manager.sources.Source
+import app.revanced.manager.compose.domain.sources.RemoteSource
+import app.revanced.manager.compose.domain.sources.LocalSource
+import app.revanced.manager.compose.domain.sources.Source
 import app.revanced.manager.compose.util.tag
 import io.ktor.http.*
 import kotlinx.coroutines.Dispatchers
@@ -17,7 +17,7 @@ import kotlinx.coroutines.withContext
 import java.io.File
 import java.io.InputStream
 
-class SourcesProvider(app: Application, private val configRepository: SourceConfigRepository) {
+class SourceRepository(app: Application, private val persistenceRepo: SourcePersistenceRepository) {
     private val sourcesDir = app.dataDir.resolve("sources").also { it.mkdirs() }
 
     private fun directoryOf(uid: Int) = sourcesDir.resolve(uid.toString())
@@ -33,7 +33,7 @@ class SourcesProvider(app: Application, private val configRepository: SourceConf
     }
 
     suspend fun loadSources() = withContext(Dispatchers.Default) {
-        val sourcesConfig = configRepository.loadConfiguration().onEach {
+        val sourcesConfig = persistenceRepo.loadConfiguration().onEach {
             Log.d(tag, "Source: $it")
         }
 
@@ -48,7 +48,7 @@ class SourcesProvider(app: Application, private val configRepository: SourceConf
     }
 
     suspend fun resetConfig() = withContext(Dispatchers.Default) {
-        configRepository.clear()
+        persistenceRepo.clear()
         _sources.emit(emptyMap())
         sourcesDir.apply {
             delete()
@@ -59,7 +59,7 @@ class SourcesProvider(app: Application, private val configRepository: SourceConf
     }
 
     suspend fun remove(source: Source) = withContext(Dispatchers.Default) {
-        configRepository.delete(source.id)
+        persistenceRepo.delete(source.id)
         sourcesDir.resolve(source.id.toString()).delete()
 
         _sources.update {
@@ -73,7 +73,7 @@ class SourcesProvider(app: Application, private val configRepository: SourceConf
         _sources.update { it.toMutableMap().apply { put(name, source) } }
 
     suspend fun createLocalSource(name: String, patches: InputStream, integrations: InputStream?) {
-        val id = configRepository.create(name, SourceLocation.Local)
+        val id = persistenceRepo.create(name, SourceLocation.Local)
         val source = LocalSource(id, directoryOf(id).create())
 
         addSource(name, source)
@@ -82,7 +82,7 @@ class SourcesProvider(app: Application, private val configRepository: SourceConf
     }
 
     suspend fun createRemoteSource(name: String, apiUrl: Url) {
-        val id = configRepository.create(name, SourceLocation.Remote(apiUrl))
+        val id = persistenceRepo.create(name, SourceLocation.Remote(apiUrl))
         addSource(name, RemoteSource(id, directoryOf(id).create()))
     }
